% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endian.R
\name{aperm_array_to_vector}
\alias{aperm_array_to_vector}
\title{Permute an R array to a linear vector of data}
\usage{
aperm_array_to_vector(x, dst, flipy = FALSE)
}
\arguments{
\item{x}{array}

\item{dst}{Specification of destination dimensions in the order of presentation
in the source data. Character vector which contains 3 strings:
'planes', 'rows', 'cols'.  The order of these strings determines
the order of output in the linear data. Currently, "planes" must always
be the final element.}

\item{flipy}{flip the array vertically. Default: FALSE}
}
\value{
vector
}
\description{
Permute an R array to a linear vector of data
}
\examples{
arr <- array(rep(c('r', 'g', 'b'), each = 6), c(2, 3, 3))
arr
aperm_array_to_vector(arr, dst = c('planes', 'cols', 'rows'))
}
\seealso{
Other data permutation functions: 
\code{\link{aperm_vector_to_array}()},
\code{\link{flip_endian}()}
}
\concept{data permutation functions}
