% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endian.R
\name{flip_endiann}
\alias{flip_endiann}
\title{Swizzle the elements in a vector}
\usage{
flip_endiann(x, size)
}
\arguments{
\item{x}{vector to swizzle. Usuually a raw vector, but can be any type}

\item{size}{swizzle size}
}
\value{
vector of the same type as the initial vector with the values
          swizzled to flip endianness.
}
\description{
This will create a new vector with the values reversed within the block size.
This is useful for changing the endianness of a set of values
}
\examples{
vec <- c(1, 2, 3, 4)
flip_endiann(vec, 2)
}
