% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-int.R
\name{read_uint8}
\alias{read_uint8}
\alias{read_int8}
\alias{read_int16}
\alias{read_uint16}
\alias{read_int32}
\alias{read_uint32}
\alias{read_int64}
\alias{read_uint64}
\title{Read integer data from a connection}
\usage{
read_uint8(con, n = 1, endian = NULL)

read_int8(con, n = 1, endian = NULL)

read_int16(con, n = 1, endian = NULL)

read_uint16(con, n = 1, endian = NULL)

read_int32(con, n = 1, endian = NULL)

read_uint32(con, n = 1, endian = NULL, promote = NULL)

read_int64(con, n = 1, endian = NULL, promote = NULL)

read_uint64(con, n = 1, endian = NULL, promote = NULL)
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{n}{Number of elements to read. Default: 1}

\item{endian}{Ordering of bytes within the file when reading multi-byte values.
Possible values: 'big' or 'little'.  
Default: NULL indicates that
endian option should be retrieved from the connection object if possible
(where the user has used \code{set_endian()}) or otherwise 
will be set to \code{"little"}}

\item{promote}{How should integer types be promoted when the type contains
values larger than R's signed integer type. Possible options 'dbl', 'raw', 'bitstring'.
Default: NULL indicates that
this option should be retrieved from the connection object if possible
(where the user has used \code{set_integer_promotion()}) or otherwise 
will be set to \code{"dbl"}.
\describe{
  \item{\code{dbl}}{Read integer values as double precision floating point}
  \item{\code{raw}}{Read integer value as a sequence of raw bytes}
  \item{\code{bitstring}}{Read integer value as a string containing 
        only the characters "0" and "1"}
}}
}
\value{
Integer data. Usually in standard R integer vector but depending on 
        the \code{promote} option may be returned in alternate formats
}
\description{
Read integer values into a standard R vector of integers or alternate
containers for large types
}
\details{
\describe{
  \item{8-bit integers}{\code{read_int8()} and \code{read_uint8()}}
  \item{16-bit integers}{\code{read_int16()} and \code{read_uint16()}}
  \item{32-bit integers}{\code{read_int32()} and \code{read_uint32()}}
  \item{64-bit integers}{\code{read_int64()} and \code{read_uint64()}}
}
}
\examples{
# Raw vector with 16 bytes (128 bits) of dummy data
data <- as.raw(1:16)
con <- rawConnection(data, 'rb')
read_uint8(con, n = 4)
read_uint64(con, n = 1)
close(con)
}
