% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_endian}
\alias{set_endian}
\title{Tag a connection with the preferred endianness}
\usage{
set_endian(con, endian = "little")
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{endian}{Default endianness to assign to this connection. 
Default: little}
}
\value{
Modified connection object
}
\description{
Tag a connection with the preferred endianness
}
\examples{
con <- textConnection("hello")
con <- set_endian("big")
}
