% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_eof_check}
\alias{set_eof_check}
\title{Set EOF handling for this connection}
\usage{
set_eof_check(con, eof_check = "error")
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{eof_check}{ignore, warn, error.  If set to 'warn' then when EOF reached
any further reads will spark a warning only.}
}
\value{
Modified connection object
}
\description{
Set EOF handling for this connection
}
\examples{
con <- textConnection("hello")
con <- set_eof_check("warn")
}
