% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_integer_promotion}
\alias{set_integer_promotion}
\title{Tag a connection with the preferred integer promotion method}
\usage{
set_integer_promotion(con, promote = "dbl")
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{promote}{Default method of promotion for uint32, uint64 and int64.
One of: dbl, raw, bitstring. Default: 'dbl'}
}
\value{
Modified connection object
}
\description{
Tag a connection with the preferred integer promotion method
}
\examples{
con <- textConnection("hello")
con <- set_integer_promotion("big")
}
