% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_na_check}
\alias{set_na_check}
\title{Check bounds on values about to be written fit into given type}
\usage{
set_na_check(con, na_check)
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{na_check}{one of: ignore, warn, error}
}
\value{
Modified connection object
}
\description{
Check bounds on values about to be written fit into given type
}
