% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-utf8.R
\name{write_raw}
\alias{write_raw}
\title{Write raw bytes}
\usage{
write_raw(con, x)
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{x}{vector to write}
}
\value{
The original connection is returned invisibly.
}
\description{
Write raw bytes
}
\examples{
con <- file(tempfile(), "wb")
write_raw(con, as.raw(1:4))
close(con)
}
