% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-utf8.R
\name{write_utf8}
\alias{write_utf8}
\alias{write_utf8_raw}
\title{Write UTF8 string}
\usage{
write_utf8(con, x)

write_utf8_raw(con, x)
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{x}{single character string}
}
\value{
The original connection is returned invisibly.
}
\description{
Write UTF8 string
}
\examples{
con <- file(tempfile(), "wb")
write_utf8(con, "hello")
close(con)
}
