% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endian.R
\name{aperm_array_to_vector}
\alias{aperm_array_to_vector}
\title{Swizzle an R array to a linear vector of data}
\usage{
aperm_array_to_vector(x, dst, flipy = FALSE)
}
\arguments{
\item{x}{array}

\item{dst}{Specification of destination dimensions in the order of presentation
in the source data. Character vector which contains 3 strings:
'planes', 'rows', 'cols'.  The order of these strings determines
the order of output in the linear data. Currently, "planes" must always
be the final element.}

\item{flipy}{flip the array vertically. Default: FALSE}
}
\value{
vector
}
\description{
Swizzle an R array to a linear vector of data
}
\examples{
mat <- matrix(c(
  'r0', 'g0', 'b0',   'r1', 'g1', 'b1',   'r2', 'g2', 'b2',   
  'r3', 'g3', 'b3',   'r4', 'g4', 'b4',   'r5', 'g5', 'b5'
), nrow = 2, ncol = 9,  byrow = TRUE)
mat
aperm_array_to_vector(mat, dst = c('rows', 'cols', 'planes'))
}
\seealso{
Other data permutation functions: 
\code{\link{aperm_vector_to_array}()},
\code{\link{flip_endian}()}
}
\concept{data permutation functions}
