% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-utf.R
\name{read_raw}
\alias{read_raw}
\title{Read raw bytes}
\usage{
read_raw(con, n = 1)
}
\arguments{
\item{con}{Connection object or raw vector. Connection objects can be 
created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{n}{Number of elements to read. Default: 1}
}
\value{
raw vector
}
\description{
Read raw bytes
}
\examples{
con <- rawConnection(charToRaw("hello12.3"))
read_raw(con, 5)
close(con)
}
\seealso{
Other data input functions: 
\code{\link{read_f64}()},
\code{\link{read_hex}()},
\code{\link{read_str}()},
\code{\link{read_uint8}()},
\code{\link{scan_dbl}()}
}
\concept{data input functions}
