% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_bounds_check}
\alias{set_bounds_check}
\title{Check bounds on values about to be written fit into given type}
\usage{
set_bounds_check(con, bounds_check = "error")
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{bounds_check}{Default bounds checking behaviour. One of: ignore, warn, error.
Default: error}
}
\value{
Modified connection object
}
\description{
Check bounds on values about to be written fit into given type
}
\examples{
con <- textConnection("hello")
con <- set_bounds_check("warn")
}
