% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_integer_promotion}
\alias{set_integer_promotion}
\title{Tag a connection with the preferred integer promotion method for types
larger that R's integer type i.e. uint32, uint64, int64}
\usage{
set_integer_promotion(con, promote = "dbl")
}
\arguments{
\item{con}{Connection object created with \code{file()}, \code{url()}, 
\code{rawConnection()} or any of the other many connection creation
functions.}

\item{promote}{Default method of promotion for uint32, uint64 and int64.
One of: 'dbl', 'hex', 'raw'. Default: 'dbl'
\describe{
  \item{\code{dbl}}{Read in integers as doubles. Integer values above 2^53
  will lose precision}
  \item{\code{hex}}{Each integer is returned as a 16-character 
  hexadecimal string}
  \item{\code{raw}}{A single raw vector containing all the integers 
  in their original form}
}}
}
\value{
Modified connection object
}
\description{
Tag a connection with the preferred integer promotion method for types
larger that R's integer type i.e. uint32, uint64, int64
}
\examples{
con <- textConnection("hello")
con <- set_integer_promotion("big")
}
